docs: Update benchmarks
Why are they so much slower than the previous benchmarks? No clue. I went back
and ran the benchmarks on the 0.2.0 release and the values are approximately
what were seeing now, and Im not keen on improving the performance of the
existing parser until we switch parsing libraries (which will be a breaking
change)